+++ src/org/jmol/adapter/readers/more/JcampdxReader.java
@@ -223,16 +223,16 @@ public class JcampdxReader extends MolReader implements JmolJDXMOLReader {
             setBonding(model, ibase);
         }
       }
-      if (!Float.isNaN(vibScale)) {
+      if (!Double.isNaN(vibScale)) {
         Logger.info("JcampdxReader applying vibration scaling of " + vibScale + " to "
             + model.ac + " atoms");
         Atom[] atoms = model.atoms;
         for (int i = model.ac; --i >= 0;) {
-          if (atoms[i].vib != null && !Float.isNaN(atoms[i].vib.z))
+          if (atoms[i].vib != null && !Double.isNaN(atoms[i].vib.z))
             atoms[i].vib.scale(vibScale);
         }
       }
-      if (!Float.isNaN(modelScale)) {
+      if (!Double.isNaN(modelScale)) {
         Logger.info("JcampdxReader applying model scaling of " + modelScale + " to "
             + model.ac + " atoms");
         Atom[] atoms = model.atoms;
